+++ src/input_processor/read_custom_tokens.rs
@@ -1,14 +1,14 @@
 use crate::{goblin_error::GoblinError, require, types::Address};
 
 pub fn read_custom_tokens<'a>(
-    custom_token_count: u8,
+    custom_token_count: usize,
     input: &'a [u8; 512],
     len: usize,
     offset: &mut usize,
 ) -> Result<&'a [Address], GoblinError> {
     let start_index = *offset;
 
-    let list_len = custom_token_count as usize * 20;
+    let list_len = custom_token_count * 20;
     *offset += list_len;
 
     require!(len >= *offset, GoblinError::InvalidPayload);
@@ -16,7 +16,7 @@ pub fn read_custom_tokens<'a>(
     let custom_tokens = unsafe {
         core::slice::from_raw_parts(
             input[start_index..*offset].as_ptr() as *const [u8; 20],
-            custom_token_count as usize,
+            custom_token_count,
         )
     };
